+++ src/layouts/RootLayout.tsx
@@ -1,7 +1,7 @@
 import { Fragment, useContext } from 'react';
 import { AuthContext } from '../store/auth-context';
 import { Navigate, Outlet } from 'react-router-dom';
-import RootNavigation from '../components/UI/RootNavigation';
+import RootNavigation from '../components/UI/navbars/RootNavigation';
 import classes from './RootLayout.module.css';
 
 const RootLayout = (): JSX.Element => {
@@ -9,7 +9,7 @@ const RootLayout = (): JSX.Element => {
 
   const route = sessionStorage.getItem('route');
 
-  if (user && !route)
+  if (user && (!route || route === '/login' || route === '/register'))
     return <Navigate to={`/${user.userRole.toLowerCase()}`} />;
   if (user && route) return <Navigate to={route} />;
