+++ src/layouts/UserLayout.tsx
@@ -2,7 +2,7 @@ import { Fragment, useContext } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
 import { AuthContext } from '../store/auth-context';
 import { UserRoles } from '../types/enums';
-import UserNavigation from '../components/UI/UserNavigation';
+import UserNavigation from '../components/UI/navbars/UserNavigation';
 import classes from './UserLayout.module.css';
 
 const UserLayout = (): JSX.Element => {
